Renaming folders for better consistency
