+++ code/src/main/java/shows/Character.java
@@ -1,4 +1,4 @@
-package characters;
+package shows;
 
 import java.util.ArrayList;
